Renamed CharacterDTO to Character Edit Response and moved to responses folder. Added Faction Id to said response
